Moving ResourceUtilities and ErrorUtilities to Framework project. Removing unnecessary AssemblyResources. Adapting MSBuildTaskHost.
