+++ src/Framework/Resources/ResourceUtilities.cs
